Update authorization. Add cross-site request forgery protection. Clients can view and update only their own data, managers can update and view everyone's data excluding admins. Managers can also manage drugs and view all pets in clinic. Admins can do everything. Add logout menu option. Add custom access denied page. Add spring mvc interceptor to inject "me" parameter with currently authenticated user name. Update all tests to use spring security mocking with admin user. Many other minor improvements.

+++ src/main/java/ru/lightstar/clinic/controller/DeleteClient.java
@@ -9,8 +9,10 @@ import org.springframework.web.servlet.mvc.support.RedirectAttributes;
 import ru.lightstar.clinic.ClinicService;
 import ru.lightstar.clinic.DrugService;
 import ru.lightstar.clinic.exception.ServiceException;
+import ru.lightstar.clinic.model.Client;
 import ru.lightstar.clinic.persistence.MessageService;
 import ru.lightstar.clinic.persistence.RoleService;
+import ru.lightstar.clinic.security.SecurityUtil;
 
 import javax.servlet.http.HttpServletRequest;
 
@@ -43,9 +45,13 @@ public class DeleteClient extends ClinicController {
      * @throws ServiceException thrown if can't delete client.
      */
     @RequestMapping(method = RequestMethod.POST)
-    public String deleteClient(@RequestParam("name") final String name, final RedirectAttributes redirectAttributes,
+    public String deleteClient(@RequestParam final String name, final RedirectAttributes redirectAttributes,
                                final HttpServletRequest request)
             throws ServiceException {
+        if (name.equals(SecurityUtil.getAuthName()) || this.getClientFromNameParam(name) == Client.NONE) {
+            return "redirect:/";
+        }
+
         this.clinicService.deleteClient(name);
         this.setMessage(redirectAttributes,  "Client deleted");
         return this.redirectToForm(request);
